#13 Merge module (comparison → comparable)
