+++ ndtools/__init__.py
@@ -11,7 +11,6 @@ __all__ = [
     "TotalOrdering",
     "collections",
     "comparable",
-    "comparison",
     "operators",
 ]
 __version__ = "0.3.0"
@@ -20,8 +19,6 @@ __version__ = "0.3.0"
 # dependencies
 from . import collections
 from . import comparable
-from . import comparison
 from . import operators
 from .collections import *
 from .comparable import *
-from .comparison import *
